# docs/contact-center-analytics/openapi.yaml
openapi: 3.0.3
info:
  title: API расширенной аналитики контакт-центра
  version: 1.0.0
  description: |
    API для загрузки обращений (ingestion) и получения агрегатов (analytics)
    для UI "Расширенная аналитика контакт-центра".
servers:
  - url: /
    description: Текущий хост (Vercel/локальный)
...
paths:
...
  /api/dictionaries/TicketSubjectOut:
    get:
      tags:
        - Dictionaries (cc_replica)
      summary: Тематики обращений исходящих звонков
      description: Dictionary backed by cc_replica."TicketSubjectOut".
      operationId: getDictionaryTicketSubjectOut
      parameters:
        - in: query
          name: q
          schema: { type: string }
          required: false
          description: Search by name.
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
          required: false
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
          required: false
        - in: query
          name: debug
          schema: { type: string, enum: ["0","1"] }
          required: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [id, code, nameRu]
                      properties:
                        id: { type: string }
                        code: { type: string }
                        nameRu: { type: string }
                        active: { type: boolean }
                  limit: { type: integer }
                  offset: { type: integer }
        "500":
          description: Database error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/dictionaries/TicketSubject:
    get:
      tags: [Dictionaries (cc_replica)]
      summary: Тематики обращений входящих звонков
      operationId: getDictionaryTicketSubject
      parameters:
        - in: query
          name: all
          schema: { type: string, enum: ["0","1"] }
          required: false
        - in: query
          name: debug
          schema: { type: string, enum: ["0","1"] }
          required: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [id, code, nameRu]
                      properties:
                        id: { type: string }
                        code: { type: string }
                        nameRu: { type: string }
                        active: { type: boolean }
        "500":
          description: Database error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: BAD_REQUEST }
            message: { type: string, example: Некорректные параметры запроса }
            details:
              type: array
              items:
                type: string